feat: add support for email triggers
Added support for email triggers. The `createEmailHandler` function is used to create an email handler that can be used to instrument email triggers. The `isEmailMessage` function is used to check if a trigger is an email message. The `instrument` function has been updated to support email handlers.

Marked `cloudflare:email` as external